refactor: 燃料翻訳ロジックをFireEquipmentOrderモデルに統合し、FuelTranslatable concernを削除
