+++ api/app/models/fire_equipment_order.rb
@@ -1,7 +1,19 @@
 class FireEquipmentOrder < ApplicationRecord
-  include FuelTranslatable
-
   belongs_to :group
   # enum gas_bottle: ガスボンベ、lp_gas: LPガス、charcoal: 炭
   enum fuel: { gas_bottle: 1, lp_gas: 2, charcoal: 3 }
+
+  FUEL_TRANSLATIONS = {
+    'gas_bottle' => 'ガスボンベ',
+    'lp_gas' => 'LPガス',
+    'charcoal' => '炭'
+  }.freeze
+
+  def fuel_japanese
+    FUEL_TRANSLATIONS[self.fuel] || self.fuel.to_s
+  end
+
+  def self.fuel_options_for_select
+    FUEL_TRANSLATIONS.map { |key, value| [value, key] }
+  end
 end
